fix(ci): unblock the cross-language conformance suite -- three independent breakages
The conformance suite itself has not RUN for many pushes.
`standard-conformance-testing` needs all five per-language jobs green, so it
has been SKIPPED and `conformance-results` then reports failure. Three
unrelated defects were holding the gate shut, and each one MASKED the ones
behind it -- a step that fails aborts the steps after it, so the first
failure in a job is the only one anybody sees.

1. THE LINT JOB, and the five gates hiding behind it.

`ruff check` fails on an unused `from pathlib import Path` in
tests/test_conformance_round_trip.py (added in bcffb7e, never used). That is
step 9 of a 16-step job, so `ruff format`, `cargo fmt`, `cargo clippy`,
eslint/prettier, gofmt, golangci-lint and the version-sync check have not run
in CI for 40+ runs. Removing the import exposes what accumulated behind it,
all of it pre-existing and all of it now fixed:

  * `ruff format` (pinned 0.16.4) -- six files, formatting only.
  * `cargo fmt --all --check` -- 27 files, formatting only. The four diffs in
    the CLI are present in the pre-rename file too; the rename below did not
    introduce them.
  * eslint -- one error, an unused `name` binding in
    src/data-source-urls.test.ts (the loop reads only `entry`, so it is
    `Object.values` now).
  * prettier -- src/join-syms.test.ts.

gofmt, golangci-lint and check-version-sync.sh were already clean.

2. THE WASM BUILD -- `no field const_arrays on type ProblemOptions`.

`ProblemOptions::const_arrays` is `cfg(not(target_arch = "wasm32"))`, but the
F32 rounding pass added in 0bbe895 reads it unguarded, so
`cargo build --target wasm32-unknown-unknown --features wasm` fails to compile
the LIB. The rounding loop is now gated like the field it reads.

Behind that first error sat a second, bigger one: the `esm` CLI has never
compiled for wasm32 AT ALL -- 11 errors, all of them the CLI reaching for the
build half of `ProblemOptions` (`build_pipeline`, `build_providers`,
`PrepareProvider`, `run_pde_tests_filtered`) that does not exist on that
target. Cargo builds every declared bin for whatever `--target` it is handed,
and builds bins even for `cargo test --test <name>`, so this broke the WASM
test suite step too, not just the build step -- the gate cannot live in the
workflow. `src/bin/esm.rs` is now a thin shim: the implementation moved to
`esm_impl.rs` and is compiled only for native targets, with an empty `main` on
wasm32. `autobins = false` (every bin is already declared explicitly) keeps
`esm_impl.rs` from being discovered as a second binary.

3. THE RUST TEST MATRIX -- five E0063s in tests/performance.rs.

`Species.default_units`, `Parameter.default_units`/`shape`/`distribution`/
`update` and `ReactionSystem.tolerance`/`tests`/`analyses` were added to the
types and every other test literal was updated. tests/performance.rs was
missed because its literals sit behind `cfg(feature = "parallel")`, which a
default `cargo test` does not compile -- only CI's `--all-features` (stable,
beta) and the MSRV leg's explicit feature list reach them. Fields added
explicitly, matching the style the other test files use.

Verified locally: ruff check/format clean, eslint 0 errors, prettier clean,
gofmt clean, version-sync clean, cargo fmt clean; the wasm32 lib builds and
the wasm test suite links; `cargo test --features <the MSRV leg's list> --test
performance` is 25 passed / 0 failed; the affected Python (145 passed) and
TypeScript (10 passed) tests still pass.
