+++ pkg/earthsci-ast-py/tests/test_conformance_round_trip.py
@@ -22,7 +22,6 @@ from __future__ import annotations
 
 import json
 from collections.abc import Iterator
-from pathlib import Path
 from typing import Any
 
 import pytest
