+++ pkg/earthsci-ast-rs/tests/performance.rs
@@ -526,6 +526,7 @@ fn test_parallel_stoichiometric_matrix_computation() {
         Species {
             units: Some("mol".to_string()),
             default: Some(1.0),
+            default_units: None,
             description: None,
             constant: None,
         },
@@ -535,6 +536,7 @@ fn test_parallel_stoichiometric_matrix_computation() {
         Species {
             units: Some("mol".to_string()),
             default: Some(1.0),
+            default_units: None,
             description: None,
             constant: None,
         },
@@ -544,6 +546,7 @@ fn test_parallel_stoichiometric_matrix_computation() {
         Species {
             units: Some("mol".to_string()),
             default: Some(0.0),
+            default_units: None,
             description: None,
             constant: None,
         },
@@ -577,13 +580,20 @@ fn test_parallel_stoichiometric_matrix_computation() {
         Parameter {
             units: Some("1/s".to_string()),
             default: Some(1.0),
+            default_units: None,
             description: None,
+            shape: None,
+            distribution: None,
+            update: None,
         },
     );
 
     let system = ReactionSystem {
         subsystems: None,
         reference: None,
+        tolerance: None,
+        tests: None,
+        analyses: None,
         species,
         parameters,
         reactions,
